IPC/COW fork fixes (#27)
* This PR started as a fix for yield_syntax_spec.rb but uncovered an issue in the process related to dynamic routing.

The core problem is that disconnected stages -- e.g. consumers with no upstream--will automatically shutdown instantly based on logic we added. See worker.rb line 60 (# If no upstream sources, this stage is disconnected)

If we remove this logic, these disconnected stages will never shut-down, unless they get a dynamic producer to connect to them (or their pipeline sends and send of stage?)

The solution is something like (A) allowing stages to be declared as disconnected: true, and (B) adding various keep-alive options for stages so they have a change to receive dynamic inputs before shutting down. A 3rd option (C) is to send Nonce signal (StageConnected) and/or KeepAlive signal to the stage initially, to declare the dynamic linkage. That might require some keepalive.

The yield example should be refactored to use connected stages

* Fix IPC and COW cases

* WIP on fixes

* Fix tests

* Cleanups

* Skip failing tests on Windows

* Remove await on fork stages